Toast雏形

+++ BadMcen Launcher/Models/CreateOrUse/CreateOrUseFiles.cs
@@ -1,5 +1,7 @@
 ﻿using BadMcen_Launcher.Views.Home.Settings.MinecraftSettings.Routine;
 using Microsoft.UI.Xaml.Controls;
+using Microsoft.Windows.AppNotifications.Builder;
+using Microsoft.Windows.AppNotifications;
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
@@ -10,6 +12,8 @@ using System.Text.Json;
 using System.Threading.Tasks;
 using Windows.ApplicationModel.Email;
 using static BadMcen_Launcher.Models.Definition;
+using Windows.System;
+using Microsoft.Toolkit.Uwp.Notifications;
 
 namespace BadMcen_Launcher.Models.CreateOrUse
 {
@@ -25,8 +29,13 @@ namespace BadMcen_Launcher.Models.CreateOrUse
         //Create json
         public void CreateJson()
         {
-            if (!File.Exists(MCPath)) File.Create(MCPath);
-            if (!File.Exists(LaunchConfigPath)) File.Create(LaunchConfigPath);
+            if (Directory.Exists(Path.Combine(pathCode.AppDataFolderPath, @"BadMC\BadMcen Launcher\Config\Settings")))
+            {
+                if (!File.Exists(MCPath)) File.Create(MCPath);
+                if (!File.Exists(LaunchConfigPath)) File.Create(LaunchConfigPath);
+            }
+           
+            
         }
         //Use files
         public class SetVersionPathJson
@@ -49,7 +58,7 @@ namespace BadMcen_Launcher.Models.CreateOrUse
                 //String to json
                 string json = JsonSerializer.Serialize(MinecraftPath);
                 //Write to json
-                await File.WriteAllTextAsync(Path.Combine(pathCode.AppDataFolderPath, @"BadMC\BadMcen Launcher\Config\Settings\MCPath.json"), json);
+                await File.WriteAllTextAsync(MCPath, json);
             }
             public void ReadJson()
             {
@@ -57,9 +66,8 @@ namespace BadMcen_Launcher.Models.CreateOrUse
                 string json = File.ReadAllText(MCPath);
                 if (!string.IsNullOrWhiteSpace(json))
                 {
-
                     List<string> JsonToList = JsonSerializer.Deserialize<List<string>>(json);
-
+                    MinecraftPath.Clear();
                     foreach (string item in JsonToList)
                     {
                         SetVersionPathPage.Instance.SetVersionPathListView.Items.Add(item);
@@ -83,31 +91,35 @@ namespace BadMcen_Launcher.Models.CreateOrUse
                 //String to json
                 string json = JsonSerializer.Serialize(config, JsonOptions);
                 //Write to json
-                await File.WriteAllTextAsync(Path.Combine(pathCode.AppDataFolderPath, @"BadMC\BadMcen Launcher\Config\LaunchConfig.json"), json);
+                await File.WriteAllTextAsync(LaunchConfigPath, json);
                 
             }
             //Read json
             public static object ReadJson(string DictionaryKey)
             {
                 //Read json and deserialize it.
-                Dictionary<string, object> dict = JsonSerializer.Deserialize<Dictionary<string, object>>(File.ReadAllText(LaunchConfigPath));
-                if (!string.IsNullOrWhiteSpace(DictionaryKey))
-                {
+                
                     try
                     {
-                        //Get the value of a specific key
-                        object ReturnValue = dict[DictionaryKey];
-                        //Return to Dictionary
-                        return ReturnValue;
+                        Dictionary<string, object> dict = JsonSerializer.Deserialize<Dictionary<string, object>>(File.ReadAllText(LaunchConfigPath));
+                        if (!string.IsNullOrWhiteSpace(DictionaryKey))
+                        {
+                            //Get the value of a specific key
+                            object ReturnValue = dict[DictionaryKey];
+                            //Return to Dictionary
+                            return ReturnValue;
+                        }
                     }
-                    catch (System.Collections.Generic.KeyNotFoundException)
+                    catch
                     {
-                        Debug.WriteLine("谁让你乱改Json的！");
-                    }
-                    
+                    new ToastContentBuilder()
+                    .AddArgument("action", "viewConversation")
+                    .AddArgument("conversationId", 9813)
+                    .AddText("Andrew sent you a picture")
+                    .AddText("Check this out, The Enchantments in Washington!")
+                    .Show();
                 }
                 return null;
-                
             }
             
         }
